simplify classes
